+++ lib/gas_adjuster/src/lib.rs
@@ -1,18 +1,13 @@
 //! This module determines the fees to pay in txs containing blocks submitted to the L1.
 
+use crate::statistics::GasStatistics;
 use alloy::providers::{DynProvider, Provider};
 use metrics::METRICS;
 use std::time::Duration;
-use std::{
-    collections::VecDeque,
-    sync::{Arc, RwLock},
-};
 use tokio::sync::watch;
 
 mod metrics;
-mod provider;
-#[cfg(test)]
-mod tests;
+mod statistics;
 
 /// This component keeps track of the median `base_fee` from the last `max_base_fee_samples` blocks.
 ///
@@ -80,14 +75,14 @@ impl GasAdjuster {
             pubdata_price_sender,
         };
         this.pubdata_price_sender
-            .send_replace(Some(this.pubdata_price_inner()));
+            .send_replace(Some(this.pubdata_price()));
 
         Ok(this)
     }
 
     /// Performs an actualization routine for `GasAdjuster`.
     /// This method is intended to be invoked periodically.
-    pub async fn keep_updated(&self) -> anyhow::Result<()> {
+    pub async fn update_fees(&mut self) -> anyhow::Result<()> {
         // Subtracting 1 from the "latest" block number to prevent errors in case
         // the info about the latest block is not yet present on the node.
         // This sometimes happens on Infura.
@@ -142,16 +137,16 @@ impl GasAdjuster {
             }
 
             self.pubdata_price_sender
-                .send_replace(Some(self.pubdata_price_inner()));
+                .send_replace(Some(self.pubdata_price()));
         }
         Ok(())
     }
 
-    pub async fn run(self: Arc<Self>) -> anyhow::Result<()> {
+    pub async fn run(mut self) -> anyhow::Result<()> {
         let mut timer = tokio::time::interval(self.config.poll_period);
         let mut attempts_failed_in_a_row = 0usize;
         loop {
-            if let Err(err) = self.keep_updated().await {
+            if let Err(err) = self.update_fees().await {
                 attempts_failed_in_a_row += 1;
                 if attempts_failed_in_a_row >= 5 {
                     tracing::warn!(
@@ -166,14 +161,14 @@ impl GasAdjuster {
         }
     }
 
-    fn gas_price_inner(&self) -> u128 {
+    pub fn gas_price(&self) -> u128 {
         let median = self.base_fee_statistics.median();
         let effective_gas_price = median + self.config.max_priority_fee_per_gas;
 
         (self.config.l1_gas_pricing_multiplier * effective_gas_price as f64) as u128
     }
 
-    fn pubdata_price_inner(&self) -> u128 {
+    pub fn pubdata_price(&self) -> u128 {
         match self.config.pubdata_mode {
             PubdataMode::Blobs => {
                 const BLOB_GAS_PER_BYTE: u128 = 1; // `BYTES_PER_BLOB` = `GAS_PER_BLOB` = 2 ^ 17.
@@ -189,84 +184,66 @@ impl GasAdjuster {
                 /// Note that it is bigger than 16 to account for potential overhead.
                 const L1_GAS_PER_PUBDATA_BYTE: u128 = 17;
 
-                self.gas_price_inner()
-                    .saturating_mul(L1_GAS_PER_PUBDATA_BYTE)
+                self.gas_price().saturating_mul(L1_GAS_PER_PUBDATA_BYTE)
             }
             PubdataMode::Validium => 0,
         }
     }
-}
 
-/// Helper structure responsible for collecting the data about recent transactions,
-/// calculating the median base fee.
-#[derive(Debug, Clone, Default)]
-pub(crate) struct GasStatisticsInner<T> {
-    samples: VecDeque<T>,
-    median_cached: T,
-    max_samples: usize,
-    last_processed_block: u64,
-}
+    /// Collects the base fee history for the specified block range.
+    ///
+    /// Returns 1 value for each block in range, assuming that these blocks exist.
+    /// Will return an error if the `upto_block` is beyond the head block.
+    async fn base_fee_history(
+        provider: &DynProvider,
+        upto_block: u64,
+        block_count: u64,
+    ) -> anyhow::Result<Vec<BaseFees>> {
+        const FEE_HISTORY_MAX_REQUEST_CHUNK: usize = 1023;
 
-impl<T: Ord + Copy + Default> GasStatisticsInner<T> {
-    fn new(max_samples: usize, block: u64, fee_history: impl IntoIterator<Item = T>) -> Self {
-        let mut statistics = Self {
-            max_samples,
-            samples: VecDeque::with_capacity(max_samples),
-            median_cached: T::default(),
-            last_processed_block: 0,
-        };
+        let mut history = Vec::with_capacity(block_count as usize);
+        let from_block = upto_block.saturating_sub(block_count - 1);
 
-        statistics.add_samples(fee_history);
+        // Here we are requesting `fee_history` from blocks
+        // `[from_block; upto_block]` in chunks of size `FEE_HISTORY_MAX_REQUEST_CHUNK`
+        // starting from the oldest block.
+        for chunk_start in (from_block..=upto_block).step_by(FEE_HISTORY_MAX_REQUEST_CHUNK) {
+            let chunk_end = (chunk_start + FEE_HISTORY_MAX_REQUEST_CHUNK as u64).min(upto_block);
+            let chunk_size = chunk_end - chunk_start + 1;
 
-        Self {
-            last_processed_block: block,
-            ..statistics
+            let fee_history = provider
+                .get_fee_history(chunk_size, chunk_end.into(), &[])
+                .await?;
+
+            if fee_history.oldest_block != chunk_start {
+                anyhow::bail!(
+                    "unexpected `oldest_block`, expected: {chunk_start}, got {}",
+                    fee_history.oldest_block
+                );
+            }
+
+            // We take `chunk_size` entries and drop data for the block after `chunk_end`.
+            for (base_fee_per_gas, base_fee_per_blob_gas) in fee_history
+                .base_fee_per_gas
+                .into_iter()
+                .zip(fee_history.base_fee_per_blob_gas)
+                .take(chunk_size as usize)
+            {
+                let fees = BaseFees {
+                    base_fee_per_gas,
+                    base_fee_per_blob_gas,
+                };
+                history.push(fees)
+            }
         }
-    }
 
-    fn median(&self) -> T {
-        self.median_cached
-    }
-
-    fn add_samples(&mut self, fees: impl IntoIterator<Item = T>) {
-        let old_len = self.samples.len();
-        self.samples.extend(fees);
-        let processed_blocks = self.samples.len() - old_len;
-        self.last_processed_block += processed_blocks as u64;
-
-        let extra = self.samples.len().saturating_sub(self.max_samples);
-        self.samples.drain(..extra);
-
-        let mut samples: Vec<_> = self.samples.iter().cloned().collect();
-
-        if !self.samples.is_empty() {
-            let (_, &mut median, _) = samples.select_nth_unstable(self.samples.len() / 2);
-            self.median_cached = median;
-        }
+        Ok(history)
     }
 }
 
-#[derive(Debug, Default)]
-pub(crate) struct GasStatistics<T>(RwLock<GasStatisticsInner<T>>);
-
-impl<T: Ord + Copy + Default> GasStatistics<T> {
-    pub fn new(max_samples: usize, block: u64, fee_history: impl IntoIterator<Item = T>) -> Self {
-        Self(RwLock::new(GasStatisticsInner::new(
-            max_samples,
-            block,
-            fee_history,
-        )))
-    }
-
-    pub fn median(&self) -> T {
-        self.0.read().unwrap().median()
-    }
-
-    pub fn add_samples(&self, fees: impl IntoIterator<Item = T>) {
-        self.0.write().unwrap().add_samples(fees)
-    }
-
-    pub fn last_processed_block(&self) -> u64 {
-        self.0.read().unwrap().last_processed_block
-    }
+/// Information about the base fees provided by the L1 client.
+#[derive(Debug, Clone, PartialEq, Eq, Default)]
+pub struct BaseFees {
+    pub base_fee_per_gas: u128,
+    pub base_fee_per_blob_gas: u128,
 }
